+++ src/Toolkit/Kafka.cs
@@ -1,3 +1,4 @@
+using System.Diagnostics;
 using Confluent.Kafka;
 using Toolkit.Types;
 
@@ -65,6 +66,33 @@ where TValue : class
           try
           {
             var consumeResult = this._inputs.Consumer.Consume(consumerCTS.Token);
+
+            // Not unit testable since the unit test suite will not be able to
+            // regiter an Activity listener that can capture the activity created here.
+            if (String.IsNullOrEmpty(this._inputs.TraceIdPath) == false)
+            {
+              string? msgTraceId = (string?)Utilities.GetByPath(
+                consumeResult.Message.Value, this._inputs.TraceIdPath
+              );
+
+              var activity = Logger.SetTraceIds(
+                msgTraceId ?? ActivityTraceId.CreateRandom().ToString(),
+                this._inputs.ActivityName ?? "Toolkit default activity name",
+                this._inputs.ActivitySourceName ?? "Toolkit default activity source name"
+              );
+
+              if (
+                this._inputs.Logger != null && activity != null &&
+                activity.TraceId.ToString() != msgTraceId
+              )
+              {
+                this._inputs.Logger.Log(
+                  Microsoft.Extensions.Logging.LogLevel.Warning,
+                  null,
+                  $"The message received from the topic '{consumeResult.Topic}', partition '{consumeResult.Partition.Value}' and offset '{consumeResult.Offset.Value}' had an invalid value for a trace ID in the node '{this._inputs.TraceIdPath}': '{msgTraceId}'"
+                );
+              }
+            }
             handler(consumeResult, null);
           }
           catch (ConsumeException e)