+++ test/unit/Toolkit.Tests/Kafka.cs
@@ -1,3 +1,4 @@
+using System.Diagnostics;
 using Confluent.Kafka;
 using Confluent.SchemaRegistry;
 using LaunchDarkly.Sdk;
@@ -11,42 +12,48 @@ namespace Toolkit.Tests;
 public class KafkaTests : IDisposable
 {
   private readonly Mock<ISchemaRegistryClient> _schemaRegistryMock;
-  private readonly Mock<IProducer<string, string>> _producerMock;
-  private readonly Mock<IConsumer<string, string>> _consumerMock;
-  private readonly Mock<Action<DeliveryResult<string, string>?, Exception?>> _handlerProducerMock;
-  private readonly Mock<Action<ConsumeResult<string, string>?, Exception?>> _handlerConsumerMock;
+  private readonly Mock<IProducer<MyKey, MyValue>> _producerMock;
+  private readonly Mock<IConsumer<MyKey, MyValue>> _consumerMock;
+  private readonly Mock<Action<DeliveryResult<MyKey, MyValue>?, Exception?>> _handlerProducerMock;
+  private readonly Mock<Action<ConsumeResult<MyKey, MyValue>?, Exception?>> _handlerConsumerMock;
   private readonly Mock<IFeatureFlags> _ffMock;
-  private KafkaInputs<string, string> _kafkaInputs;
+  private readonly Mock<ILogger> _loggerMock;
+  private KafkaInputs<MyKey, MyValue> _kafkaInputs;
 
   public KafkaTests()
   {
     this._schemaRegistryMock = new Mock<ISchemaRegistryClient>(MockBehavior.Strict);
-    this._producerMock = new Mock<IProducer<string, string>>(MockBehavior.Strict);
-    this._consumerMock = new Mock<IConsumer<string, string>>(MockBehavior.Strict);
-    this._handlerProducerMock = new Mock<Action<DeliveryResult<string, string>?, Exception?>>(MockBehavior.Strict);
-    this._handlerConsumerMock = new Mock<Action<ConsumeResult<string, string>?, Exception?>>(MockBehavior.Strict);
+    this._producerMock = new Mock<IProducer<MyKey, MyValue>>(MockBehavior.Strict);
+    this._consumerMock = new Mock<IConsumer<MyKey, MyValue>>(MockBehavior.Strict);
+    this._handlerProducerMock = new Mock<Action<DeliveryResult<MyKey, MyValue>?, Exception?>>(MockBehavior.Strict);
+    this._handlerConsumerMock = new Mock<Action<ConsumeResult<MyKey, MyValue>?, Exception?>>(MockBehavior.Strict);
     this._ffMock = new Mock<IFeatureFlags>(MockBehavior.Strict);
+    this._loggerMock = new Mock<ILogger>(MockBehavior.Strict);
 
-    this._producerMock.Setup(s => s.ProduceAsync(It.IsAny<string>(), It.IsAny<Message<string, string>>(), It.IsAny<CancellationToken>()))
-      .Returns(Task.FromResult(new DeliveryResult<string, string> { }));
+    this._producerMock.Setup(s => s.ProduceAsync(It.IsAny<string>(), It.IsAny<Message<MyKey, MyValue>>(), It.IsAny<CancellationToken>()))
+      .Returns(Task.FromResult(new DeliveryResult<MyKey, MyValue> { }));
     this._producerMock.Setup(s => s.Flush(It.IsAny<CancellationToken>()));
 
     this._consumerMock.Setup(s => s.Subscribe(It.IsAny<IEnumerable<string>>()));
     this._consumerMock.SetupSequence(s => s.Consume(It.IsAny<CancellationToken>()))
-      .Returns(new ConsumeResult<string, string>())
+      .Returns(new ConsumeResult<MyKey, MyValue>())
       .Throws(new OperationCanceledException());
-    this._consumerMock.Setup(s => s.Commit(It.IsAny<ConsumeResult<string, string>>()));
+    this._consumerMock.Setup(s => s.Commit(It.IsAny<ConsumeResult<MyKey, MyValue>>()));
 
-    this._handlerProducerMock.Setup(s => s(It.IsAny<DeliveryResult<string, string>>(), It.IsAny<Exception?>()));
-    this._handlerConsumerMock.Setup(s => s(It.IsAny<ConsumeResult<string, string>>(), It.IsAny<Exception?>()));
+    this._handlerProducerMock.Setup(s => s(It.IsAny<DeliveryResult<MyKey, MyValue>>(), It.IsAny<Exception?>()));
+    this._handlerConsumerMock.Setup(s => s(It.IsAny<ConsumeResult<MyKey, MyValue>>(), It.IsAny<Exception?>()));
 
     this._ffMock.Setup(s => s.GetBoolFlagValue(It.IsAny<string>()))
       .Returns(true);
     this._ffMock.Setup(s => s.SubscribeToValueChanges(It.IsAny<string>(), It.IsAny<Action<FlagValueChangeEvent>?>()));
 
-    this._kafkaInputs = new KafkaInputs<string, string>
+    this._loggerMock.Setup(s => s.Log(It.IsAny<Microsoft.Extensions.Logging.LogLevel>(), It.IsAny<Exception?>(), It.IsAny<string>()));
+
+    this._kafkaInputs = new KafkaInputs<MyKey, MyValue>
     {
       SchemaRegistry = this._schemaRegistryMock.Object,
+      Logger = this._loggerMock.Object,
+      TraceIdPath = "correlationId",
     };
   }
 
@@ -58,18 +65,19 @@ public class KafkaTests : IDisposable
     this._handlerProducerMock.Reset();
     this._handlerConsumerMock.Reset();
     this._ffMock.Reset();
+    this._loggerMock.Reset();
   }
 
   [Fact]
   public void Publish_ItShouldCallProduceAsyncFromTheProducerInstanceOnceWithTheExpectedArguments()
   {
     this._kafkaInputs.Producer = this._producerMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
-    var testMessage = new Message<string, string>
+    var testMessage = new Message<MyKey, MyValue>
     {
-      Key = "test msg key",
-      Value = "test msg value"
+      Key = new MyKey { Id = "test msg key" },
+      Value = new MyValue { },
     };
     sut.Publish("test topic name", testMessage, this._handlerProducerMock.Object);
 
@@ -80,12 +88,12 @@ public class KafkaTests : IDisposable
   public void Publish_ItShouldCallFlushFromTheProducerInstanceOnceWithTheExpectedArguments()
   {
     this._kafkaInputs.Producer = this._producerMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
-    var testMessage = new Message<string, string>
+    var testMessage = new Message<MyKey, MyValue>
     {
-      Key = "test msg key",
-      Value = "test msg value"
+      Key = new MyKey { Id = "test msg key" },
+      Value = new MyValue { },
     };
     sut.Publish("test topic name", testMessage, this._handlerProducerMock.Object);
 
@@ -95,16 +103,16 @@ public class KafkaTests : IDisposable
   [Fact]
   public async Task Publish_ItShouldCallTheHandlerReceivedAsInputOnceWithTheExpectedArguments()
   {
-    var deliveryRes = new DeliveryResult<string, string> { };
-    this._producerMock.Setup(s => s.ProduceAsync(It.IsAny<string>(), It.IsAny<Message<string, string>>(), It.IsAny<CancellationToken>()))
+    var deliveryRes = new DeliveryResult<MyKey, MyValue> { };
+    this._producerMock.Setup(s => s.ProduceAsync(It.IsAny<string>(), It.IsAny<Message<MyKey, MyValue>>(), It.IsAny<CancellationToken>()))
       .Returns(Task.FromResult(deliveryRes));
     this._kafkaInputs.Producer = this._producerMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
-    var testMessage = new Message<string, string>
+    var testMessage = new Message<MyKey, MyValue>
     {
-      Key = "test msg key",
-      Value = "test msg value"
+      Key = new MyKey { Id = "test msg key" },
+      Value = new MyValue { },
     };
     sut.Publish("test topic name", testMessage, this._handlerProducerMock.Object);
     await Task.Delay(500);
@@ -115,17 +123,17 @@ public class KafkaTests : IDisposable
   [Fact]
   public async Task Publish_IfTheReturnOfCallingProduceAsyncFromTheProducerInstanceIsATaskThatDidNotCompleteSuccessfully_ItShouldCallTheHandlerReceivedAsInputOnceWithTheExpectedArguments()
   {
-    var deliveryRes = new DeliveryResult<string, string> { };
+    var deliveryRes = new DeliveryResult<MyKey, MyValue> { };
     var testEx = new Exception("test error msg");
-    this._producerMock.Setup(s => s.ProduceAsync(It.IsAny<string>(), It.IsAny<Message<string, string>>(), It.IsAny<CancellationToken>()))
-      .Returns(Task.FromException<DeliveryResult<string, string>>(testEx));
+    this._producerMock.Setup(s => s.ProduceAsync(It.IsAny<string>(), It.IsAny<Message<MyKey, MyValue>>(), It.IsAny<CancellationToken>()))
+      .Returns(Task.FromException<DeliveryResult<MyKey, MyValue>>(testEx));
     this._kafkaInputs.Producer = this._producerMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
-    var testMessage = new Message<string, string>
+    var testMessage = new Message<MyKey, MyValue>
     {
-      Key = "test msg key",
-      Value = "test msg value"
+      Key = new MyKey { Id = "test msg key" },
+      Value = new MyValue { },
     };
     sut.Publish("test topic name", testMessage, this._handlerProducerMock.Object);
     await Task.Delay(500);
@@ -136,12 +144,12 @@ public class KafkaTests : IDisposable
   [Fact]
   public void Publish_IfAProducerWasNotProvidedInTheInputs_ItShouldThrowAnException()
   {
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
-    var testMessage = new Message<string, string>
+    var testMessage = new Message<MyKey, MyValue>
     {
-      Key = "test msg key",
-      Value = "test msg value"
+      Key = new MyKey { Id = "test msg key" },
+      Value = new MyValue { },
     };
 
     var e = Assert.Throws<Exception>(() => sut.Publish("test topic name", testMessage, this._handlerProducerMock.Object));
@@ -153,7 +161,7 @@ public class KafkaTests : IDisposable
   {
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     var consumerCTS = new CancellationTokenSource();
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, consumerCTS);
@@ -169,7 +177,7 @@ public class KafkaTests : IDisposable
   {
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     var consumerCTS = new CancellationTokenSource();
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, consumerCTS);
@@ -183,14 +191,27 @@ public class KafkaTests : IDisposable
   [Fact]
   public async Task Subscribe_WithCancelToken_ItShouldCallTheHandlerReceivedAsInputAtLeastOnceWithTheExpectedArguments()
   {
-    var consumeRes = new ConsumeResult<string, string>();
+    var consumeRes = new ConsumeResult<MyKey, MyValue>
+    {
+      Message = new Message<MyKey, MyValue>
+      {
+        Value = new MyValue
+        {
+          Id = ActivityTraceId.CreateRandom().ToString(),
+        },
+      },
+      Partition = new Partition(123),
+      Offset = new Offset(987),
+      Topic = "test topic",
+    };
     this._consumerMock.SetupSequence(s => s.Consume(It.IsAny<CancellationToken>()))
       .Returns(consumeRes)
       .Throws(new OperationCanceledException());
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
+    this._kafkaInputs.TraceIdPath = "Id";
     var consumerCTS = new CancellationTokenSource();
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, consumerCTS);
@@ -204,13 +225,13 @@ public class KafkaTests : IDisposable
   [Fact]
   public async Task Subscribe_WithCancelToken_IfAConsumeExceptionIsThrown_ItShouldContinueCallingConsumeFromTheConsumerInstanceWithTheExpectedArguments()
   {
-    var consumeRes = new ConsumeResult<string, string>();
+    var consumeRes = new ConsumeResult<MyKey, MyValue>();
     this._consumerMock.Setup(s => s.Consume(It.IsAny<CancellationToken>()))
       .Throws(new ConsumeException(new ConsumeResult<byte[], byte[]>(), new Error(ErrorCode.Local_Fail), new Exception()));
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     var consumerCTS = new CancellationTokenSource();
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, consumerCTS);
@@ -222,14 +243,14 @@ public class KafkaTests : IDisposable
   [Fact]
   public async Task Subscribe_WithCancelToken_IfAConsumeExceptionIsThrown_ItShouldCallTheHandlerReceivedAsInputOnceWithTheExpectedArguments()
   {
-    var consumeRes = new ConsumeResult<string, string>();
+    var consumeRes = new ConsumeResult<MyKey, MyValue>();
     var innerEx = new Exception();
     this._consumerMock.Setup(s => s.Consume(It.IsAny<CancellationToken>()))
       .Throws(new ConsumeException(new ConsumeResult<byte[], byte[]>(), new Error(ErrorCode.Local_Fail), innerEx));
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     var consumerCTS = new CancellationTokenSource();
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, consumerCTS);
@@ -241,14 +262,14 @@ public class KafkaTests : IDisposable
   [Fact]
   public async Task Subscribe_WithCancelToken_IfAnOperationCanceledExceptionIsThrown_ItShouldCallTheHandlerReceivedAsInputOnceWithTheExpectedArguments()
   {
-    var consumeRes = new ConsumeResult<string, string>();
+    var consumeRes = new ConsumeResult<MyKey, MyValue>();
     var innerEx = new Exception();
     this._consumerMock.Setup(s => s.Consume(It.IsAny<CancellationToken>()))
       .Throws(new OperationCanceledException("exception msg from test", innerEx));
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     var consumerCTS = new CancellationTokenSource();
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, consumerCTS);
@@ -260,14 +281,14 @@ public class KafkaTests : IDisposable
   [Fact]
   public async Task Subscribe_WithCancelToken_IfAnExceptionIsThrown_ItShouldCallTheHandlerReceivedAsInputOnceWithTheExpectedArguments()
   {
-    var consumeRes = new ConsumeResult<string, string>();
+    var consumeRes = new ConsumeResult<MyKey, MyValue>();
     var innerEx = new Exception("exception msg from test");
     this._consumerMock.Setup(s => s.Consume(It.IsAny<CancellationToken>()))
       .Throws(innerEx);
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     var consumerCTS = new CancellationTokenSource();
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, consumerCTS);
@@ -279,7 +300,7 @@ public class KafkaTests : IDisposable
   [Fact]
   public void Subscribe_WithCancelToken_IfAConsumerWasNotProvidedInTheInputs_ItShouldThrowAnException()
   {
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["test topic name"];
     var e = Assert.Throws<Exception>(() => sut.Subscribe(topics, this._handlerConsumerMock.Object));
@@ -290,7 +311,7 @@ public class KafkaTests : IDisposable
   public async Task Subscribe_WithCancelToken_IfAConsumerCancellationTokenSourceWasNotProvidedInTheInputs_ItShouldCallConsumeFromTheConsumerInstanceAtLeastOnceWithTheExpectedArguments()
   {
     this._kafkaInputs.Consumer = this._consumerMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object);
@@ -304,7 +325,7 @@ public class KafkaTests : IDisposable
   {
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     this._kafkaInputs.FeatureFlags = this._ffMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, "some ff key");
@@ -318,7 +339,7 @@ public class KafkaTests : IDisposable
   {
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     this._kafkaInputs.FeatureFlags = this._ffMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, "some ff key");
@@ -332,7 +353,7 @@ public class KafkaTests : IDisposable
   {
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     this._kafkaInputs.FeatureFlags = this._ffMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, "some ff key");
@@ -346,7 +367,7 @@ public class KafkaTests : IDisposable
   {
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     this._kafkaInputs.FeatureFlags = this._ffMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, "some ff key");
@@ -358,14 +379,17 @@ public class KafkaTests : IDisposable
   [Fact]
   public async Task Subscribe_WithFFKey_ItShouldCallTheHandlerReceivedAsInputAtLeastOnceWithTheExpectedArguments()
   {
-    var consumeRes = new ConsumeResult<string, string>();
+    var consumeRes = new ConsumeResult<MyKey, MyValue>
+    {
+      Message = new Message<MyKey, MyValue> { },
+    };
     this._consumerMock.SetupSequence(s => s.Consume(It.IsAny<CancellationToken>()))
       .Returns(consumeRes)
       .Throws(new OperationCanceledException());
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     this._kafkaInputs.FeatureFlags = this._ffMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, "some ff key");
@@ -383,7 +407,7 @@ public class KafkaTests : IDisposable
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     this._kafkaInputs.FeatureFlags = this._ffMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, "some ff key");
@@ -407,7 +431,7 @@ public class KafkaTests : IDisposable
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     this._kafkaInputs.FeatureFlags = this._ffMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, "some ff key");
@@ -427,7 +451,7 @@ public class KafkaTests : IDisposable
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     this._kafkaInputs.FeatureFlags = this._ffMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, "some ff key");
@@ -444,7 +468,7 @@ public class KafkaTests : IDisposable
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     this._kafkaInputs.FeatureFlags = this._ffMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, "some ff key");
@@ -457,7 +481,7 @@ public class KafkaTests : IDisposable
   public void Subscribe_WithFFKey_IfAFeatureFlagInstanceWasNotProvidedInTheInputs_ItShouldThrowAnException()
   {
     this._kafkaInputs.Consumer = this._consumerMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
 
@@ -468,14 +492,14 @@ public class KafkaTests : IDisposable
   [Fact]
   public async Task Subscribe_WithFFKey_IfAConsumeExceptionIsThrown_ItShouldCallTheHandlerReceivedAsInputOnceWithTheExpectedArguments()
   {
-    var consumeRes = new ConsumeResult<string, string>();
+    var consumeRes = new ConsumeResult<MyKey, MyValue>();
     var innerEx = new Exception();
     this._consumerMock.Setup(s => s.Consume(It.IsAny<CancellationToken>()))
       .Throws(new ConsumeException(new ConsumeResult<byte[], byte[]>(), new Error(ErrorCode.Local_Fail), innerEx));
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     this._kafkaInputs.FeatureFlags = this._ffMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, "some ff key");
@@ -487,14 +511,14 @@ public class KafkaTests : IDisposable
   [Fact]
   public async Task Subscribe_WithFFKey_IfAnOperationCanceledExceptionIsThrown_ItShouldCallTheHandlerReceivedAsInputOnceWithTheExpectedArguments()
   {
-    var consumeRes = new ConsumeResult<string, string>();
+    var consumeRes = new ConsumeResult<MyKey, MyValue>();
     var innerEx = new Exception();
     this._consumerMock.Setup(s => s.Consume(It.IsAny<CancellationToken>()))
       .Throws(new OperationCanceledException("exception msg from test", innerEx));
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     this._kafkaInputs.FeatureFlags = this._ffMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, "some ff key");
@@ -506,14 +530,14 @@ public class KafkaTests : IDisposable
   [Fact]
   public async Task Subscribe_WithFFKey_IfAnExceptionIsThrown_ItShouldCallTheHandlerReceivedAsInputOnceWithTheExpectedArguments()
   {
-    var consumeRes = new ConsumeResult<string, string>();
+    var consumeRes = new ConsumeResult<MyKey, MyValue>();
     var innerEx = new Exception("exception msg from test");
     this._consumerMock.Setup(s => s.Consume(It.IsAny<CancellationToken>()))
       .Throws(innerEx);
 
     this._kafkaInputs.Consumer = this._consumerMock.Object;
     this._kafkaInputs.FeatureFlags = this._ffMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
     IEnumerable<string> topics = ["some other test topic name"];
     sut.Subscribe(topics, this._handlerConsumerMock.Object, "some ff key");
@@ -526,9 +550,9 @@ public class KafkaTests : IDisposable
   public void Commit_ItShouldCallCommitFromTheConsumerInstanceOnceWithTheExpectedArguments()
   {
     this._kafkaInputs.Consumer = this._consumerMock.Object;
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
-    var consumeRes = new ConsumeResult<string, string>();
+    var consumeRes = new ConsumeResult<MyKey, MyValue>();
     sut.Commit(consumeRes);
 
     this._consumerMock.Verify(m => m.Commit(consumeRes), Times.Once());
@@ -537,9 +561,27 @@ public class KafkaTests : IDisposable
   [Fact]
   public void Commit_IfAConsumerWasNotProvidedInTheInputs_ItShouldThrowAnException()
   {
-    var sut = new Kafka<string, string>(this._kafkaInputs);
+    var sut = new Kafka<MyKey, MyValue>(this._kafkaInputs);
 
-    var e = Assert.Throws<Exception>(() => sut.Commit(new ConsumeResult<string, string>()));
+    var e = Assert.Throws<Exception>(() => sut.Commit(new ConsumeResult<MyKey, MyValue>()));
     Assert.Equal("An instance of IConsumer was not provided in the inputs.", e.Message);
   }
+}
+
+public class MyKey
+{
+  public required string Id { get; set; }
+}
+
+public class MyValue
+{
+  public string? Id { get; set; }
+  public string? Name { get; set; }
+  public List<MyValueDoc>? Docs { get; set; }
+}
+
+public class MyValueDoc
+{
+  public int? Count { get; set; }
+  public string? Desc { get; set; }
 }